🦄 refactor(reserve): 拆分模块
