+++ src/func/misc.ts
@@ -1,6 +1,8 @@
-import { warn, error, debug } from "../utils";
-import * as serverApi from '../serverApi';
-import { settings } from '../global-status';
+import { openTab } from 'siyuan';
+
+import { warn, error, debug, app } from "@/utils";
+import * as serverApi from '@/serverApi';
+import { settings } from '@/global-status';
 import { getDailynoteSprig, renderDailynotePath } from './dailynote';
 
 const default_sprig = `/daily note/{{now | date "2006/01"}}/{{now | date "2006-01-02"}}`
@@ -82,4 +84,41 @@ export async function filterExistsBlocks(blockIds: string[]): Promise<Set<string
         idSet.add(block.id);
     });
     return idSet;
+}
+
+export function compareVersion(v1Str: string, v2Str: string) {
+    let v1 = v1Str.split('.')
+    let v2 = v2Str.split('.')
+    const len = Math.max(v1.length, v2.length)
+
+    while (v1.length < len) {
+        v1.push('0')
+    }
+    while (v2.length < len) {
+        v2.push('0')
+    }
+
+    for (let i = 0; i < len; i++) {
+        const num1 = parseInt(v1[i])
+        const num2 = parseInt(v2[i])
+
+        if (num1 > num2) {
+            return 1
+        } else if (num1 < num2) {
+            return -1
+        }
+    }
+
+    return 0
+}
+
+export function openBlock(blockId: BlockId) {
+    openTab({
+        app: app,
+        doc: {
+            id: blockId,
+            action: ['cb-get-focus'],
+            zoomIn: true
+        }
+    });
 }
